feat: scaffold turn flow and gates
