+++ server/src/turns/gates.ts
@@ -1,0 +1,26 @@
+import type { GateContext } from '../types';
+
+export function budgetGate(ctx: GateContext): GateContext {
+  // Placeholder: ensure slot was funded
+  return ctx;
+}
+
+export function inputsGate(ctx: GateContext): GateContext {
+  // Placeholder: check for required inputs
+  return ctx;
+}
+
+export function logisticsGate(ctx: GateContext): GateContext {
+  // Placeholder: validate logistics capacity
+  return ctx;
+}
+
+export function laborGate(ctx: GateContext): GateContext {
+  // Placeholder: assign available labor
+  return ctx;
+}
+
+export function suitabilityGate(ctx: GateContext): GateContext {
+  // Placeholder: apply site modifiers
+  return ctx;
+}
