+++ server/src/turns/plan.ts
@@ -1,0 +1,9 @@
+import type { TurnPlan } from '../types';
+
+export function createEmptyPlan(): TurnPlan {
+  return {
+    budgets: {},
+    policies: {},
+    priorities: {}
+  };
+}
